+++ convert-junit4-to-junit5/src/test/java/jb/convert/ast/GeneralMigrationTest.java
@@ -1,0 +1,16 @@
+package jb.convert.ast;
+
+import org.junit.jupiter.api.Test;
+
+import static jb.JunitConversionLogicFixture.assertAfterAddingClassAfter;
+
+class GeneralMigrationTest {
+
+    @Test
+    void importWildcardMainPackage() {
+        String originalImport = "import org.junit.*;";
+        String expectedImport = "import org.junit.jupiter.api.*;";
+        assertAfterAddingClassAfter(originalImport, expectedImport);
+    }
+
+}
